User authentication refactoring in progress.
