+++ src/main/java/com/dmmsoft/LoginServletGoogle.java
@@ -1,5 +1,7 @@
-package com.dmmsoft.loginwithgoogle;
+package com.dmmsoft;
 
+import com.dmmsoft.loginwithgoogle.IdTokenVerifierAndParser;
+import com.dmmsoft.user.User;
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdToken;
 
 import javax.servlet.ServletException;
@@ -14,7 +16,7 @@ import java.io.IOException;
  * Created by daniel on 29.05.17.
  */
 @WebServlet(urlPatterns = "login2")
-public class LoginServlet extends HttpServlet {
+public class LoginServletGoogle extends HttpServlet {
     @Override
     protected void doPost (HttpServletRequest req,
                            HttpServletResponse resp)
@@ -27,13 +29,17 @@ public class LoginServlet extends HttpServlet {
             GoogleIdToken.Payload payLoad = IdTokenVerifierAndParser.getPayload(idToken);
             String name = (String) payLoad.get("name");
             String email = payLoad.getEmail();
-            System.out.println("User name: " + name);
-            System.out.println("User email: " + email);
+
+
+
 
             HttpSession session = req.getSession(true);
             session.setAttribute("userName", name);
             req.getServletContext();
             resp.sendRedirect("userview.jsp");
+
+
+
 
         } catch (Exception e) {
             throw new RuntimeException(e);
