+++ src/main/webapp/login.jsp
@@ -77,7 +77,7 @@
         //do not post all above info to the server because that is not secure.
         //just send the id_token
 
-        var redirectUrl = 'login';
+        var redirectUrl = 'login2';
 
         //using jquery to post data dynamically
         var form = $('<form action="' + redirectUrl + '" method="post">' +
